[util/timing] fix Eventual being second-precise
This fixes a bug that would cause Eventual callables to be spammed
multiple times over a few seconds. The core issue behind this bug was
that we only used the number of seconds between two points in time as
our metric for delay. This would mean that a duration of 5 seconds and
500 milliseconds would only sleep for 5 seconds, causing the eventual to
trigger 500 milliseconds early, which could cause our parent to re-call
set_next_schedule and continue spamming until the correct datetime was
reached.

The fix here is to include the most precise unit of time we can within
our calculation, in this case microseconds.