'I'm gonna rewrite the web app' - Me, a couple days ago
And now I'm here. It's taken me a lot to learn web technologies and be comfortable enough in taking on a project like this. I'm proud of me. You should be too. Many of this stuff hasn't been changed since it was originally migrated to React by it's previous maintainer... in **2017**.

I'm also using this as a chance to give myself a random challenge to try to keep the bundle as small as possible. This includes migrating to Preact later, as well as the work that's already gone in with using Wouter and Zustand.

The following commits are mostly if not only design changes, as this is really all the frontend is at it's core.